REFACTOR and fix CLI load
