+++ core/src/main/kotlin/uk/dioxic/mgenerate/serialization/RateSerializer.kt
@@ -1,8 +1,12 @@
-package uk.dioxic.mgenerate.worker.model
+package uk.dioxic.mgenerate.serialization
 
 import kotlinx.serialization.json.JsonContentPolymorphicSerializer
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.jsonObject
+import uk.dioxic.mgenerate.worker.model.PeriodRate
+import uk.dioxic.mgenerate.worker.model.RampedRate
+import uk.dioxic.mgenerate.worker.model.Rate
+import uk.dioxic.mgenerate.worker.model.TpsRate
 
 object RateSerializer : JsonContentPolymorphicSerializer<Rate>(Rate::class) {
