+++ core/src/main/kotlin/uk/dioxic/mgenerate/serialization/TemplateSerializer.kt
@@ -1,4 +1,4 @@
-package uk.dioxic.mgenerate
+package uk.dioxic.mgenerate.serialization
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.PrimitiveKind
@@ -7,6 +7,7 @@ import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 import kotlinx.serialization.json.JsonObject
+import uk.dioxic.mgenerate.Template
 import uk.dioxic.mgenerate.Template.Companion.defaultRegistry
 import uk.dioxic.mgenerate.codecs.TemplateDocumentCodec
 import uk.dioxic.mgenerate.extensions.asJsonDecoder
