+++ core/src/main/kotlin/uk/dioxic/mgenerate/worker/model/Workload.kt
@@ -2,6 +2,7 @@ package uk.dioxic.mgenerate.worker.model
 
 import kotlinx.serialization.Serializable
 import uk.dioxic.mgenerate.Template
+import uk.dioxic.mgenerate.serialization.WorkloadSerializer
 import uk.dioxic.mgenerate.worker.Named
 import kotlin.time.Duration
 
@@ -25,7 +26,7 @@ data class RateWorkload(
     override val state: Template = Template.EMPTY,
     override val executor: Executor<*>,
     override val count: Long = 1,
-    val rate: Rate = Unlimited,
+    val rate: Rate = UnlimitedRate,
     val startDelay: Duration = Duration.ZERO,
 ) : Workload()
